requirements.txt
streamlit
pandas
numpy
pillow
tensorflow
scikit-learn
matplotlib